Add a test for IDAllocator.

+++ source/engine/src/util/IDAllocator.cpp
@@ -22,7 +22,7 @@ int IDAllocator::allocate()
 	}
 	else
 	{
-		n = m_used.size();
+		n = static_cast<int>(m_used.size());
 	}
 
 	m_used.insert(n);
